Introduce TalkJS Chatbox into each poker room
